Allow access to the user's keyring
This commit makes it possible to automatically unlock a user's GPG key.
Previously, signing commits in VSCodium could fail unexpectedly.
This issue effected users storing the GPG key's passphrase in a keyring.
Even if the keyring was unlocked, like upon login, signing failed.
However, if the GPG key was unlocked elsewhere, then signing would work.
For instance, signing a commit on the command-line would load the key.
This enabled access for VSCodium, making the issue hard to reproduce.
This commit permits access to the user's keyring, solving the problem.
GPG key passphrases available in a keyring are now accessible.
When a user goes to sign a commit, everything will just work this way.
This assumes the keyring is unlocked and the passphrase stored therein.